+++ Plugins/UtilityPlugin/ProjectDescriptionHelpers/Dependency+Project.swift
@@ -10,6 +10,11 @@ extension TargetDependency {
             public struct Domain {}
             public struct UserInterface {}
         }
+        public struct MOITShare {
+            public struct Data {}
+            public struct Domain {}
+            public struct UserInterface {}
+        }
     }
     
     public struct Core {
